airflow/requirements.txt
# Airflow Cosmos - dbt integration
astronomer-cosmos>=1.7.0

# dbt dependencies
dbt-core>=1.9.1
dbt-duckdb>=1.9.1

# Database
duckdb>=1.1.0

# Extract pipeline dependencies (from strava-extract)
dlt[duckdb]>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
psutil>=6.1.1
tqdm>=4.67.1
pendulum>=3.0.0
